# src/shorturl/web/models/key.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from .base_model_ import Model
from .. import util


class Key(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, key: str=None):  # noqa: E501
        """Key - a model defined in Swagger

        :param key: The key of this Key.  # noqa: E501
        :type key: str
        """
        self.swagger_types = {
            'key': str
        }

        self.attribute_map = {
            'key': 'key'
        }

        self._key = key

    @classmethod
    def from_dict(cls, dikt) -> 'Key':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Key of this Key.  # noqa: E501
        :rtype: Key
        """
        return util.deserialize_model(dikt, cls)

    @property
    def key(self) -> str:
        """Gets the key of this Key.


        :return: The key of this Key.
        :rtype: str
        """
        return self._key

    @key.setter
    def key(self, key: str):
        """Sets the key of this Key.


        :param key: The key of this Key.
        :type key: str
        """
        if key is None:
            raise ValueError("Invalid value for `key`, must not be `None`")  # noqa: E501

        self._key = key