# src/shorturl/web/models/update_user.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from .base_model_ import Model
from .. import util


class UpdateUser(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, password: str=None):  # noqa: E501
        """UpdateUser - a model defined in Swagger

        :param password: The password of this UpdateUser.  # noqa: E501
        :type password: str
        """
        self.swagger_types = {
            'password': str
        }

        self.attribute_map = {
            'password': 'password'
        }

        self._password = password

    @classmethod
    def from_dict(cls, dikt) -> 'UpdateUser':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UpdateUser of this UpdateUser.  # noqa: E501
        :rtype: UpdateUser
        """
        return util.deserialize_model(dikt, cls)

    @property
    def password(self) -> str:
        """Gets the password of this UpdateUser.


        :return: The password of this UpdateUser.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password: str):
        """Sets the password of this UpdateUser.


        :param password: The password of this UpdateUser.
        :type password: str
        """
        if password is None:
            raise ValueError("Invalid value for `password`, must not be `None`")  # noqa: E501

        self._password = password